Moved the maps to the correct place
